working on seeders

+++ app/Http/Controllers/DashboardController.php
@@ -50,8 +50,6 @@ class DashboardController extends Controller
         $data['email'] = Request::input('email');
         $data['role'] = Request::input('role');
         $data['company'] = Request::input('company_id')['id'];
-        // $company = Request::input('company_id')['id'];
-        $company = Company::where('id', $data['company'])->first();
 
         Request::validate([
             'email' => ['required'],
@@ -59,18 +57,32 @@ class DashboardController extends Controller
             'company_id' => ['required'],
         ]);
 
-        // dd(Request::input('email'). '-------' . Request::input('role') . '-------' . Request::input('company_id')['id']);
-
         $userexist = User::where('email',$data['email'])->first('id');
+        if(auth()->user()->email != $data['email'])
+        {
+            if($userexist){
+                $userexist->roles()->detach();
+                $userexist->assignRole($data['role']);
 
-        // dd($userexist->id);
-        if($userexist){
-            $userexist->roles()->detach();
-            $userexist->assignRole($data['role']);
-
-            $company->users()->attach($userexist->id);
-        }else{
-            return Redirect::back()->with('warning', 'User email doesn\'t exists');
+                $company = Company::where('id', $data['company'])->with('users')->first();
+                $check = true;
+                foreach($company->users as $comp_user)
+                {
+                    if($comp_user->email == $data['email'])
+                    {
+                        $check = false;
+                        break;
+                    }
+                }
+                if($check)
+                {
+                    $company->users()->attach($userexist->id);
+                }
+            } else {
+                return Redirect::back()->with('warning', 'User email doesn\'t exists');
+            }
+        } else {
+                return Redirect::back()->with('warning', 'You can\'t change your own role');
         }
 
         return Redirect::back()->with('success', 'Role assigned.');
